src/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
google-cloud-bigquery==3.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
